[chaos] Move camera hooks
