+++ src/camera/hooks/use-registering.js
@@ -1,0 +1,11 @@
+import { useEffect } from 'react'
+import { useDispatch } from 'react-redux'
+import { actions } from 'loading/slices/loading'
+
+export const useRegistering = () => {
+  const dispatch = useDispatch()
+  useEffect(() => {
+    dispatch(actions.show('registering-camera'))
+    dispatch(actions.verify({}))
+  }, [dispatch])
+}
