+++ src/camera/hooks/use-should-hide.js
@@ -1,0 +1,3 @@
+import { useSelector } from 'react-redux'
+
+export const useShouldHide = () => useSelector(state => state.loading.includes('registering-camera'))
